+++ lib/td_dd_web/controllers/metadata_controller.ex
@@ -6,6 +6,7 @@ defmodule TdDdWeb.MetadataController do
   alias TdDd.Auth.Guardian.Plug, as: GuardianPlug
   alias TdDd.DataStructures
   alias TdDd.Loader
+  alias TdDd.Search.IndexWorker
 
   @taxonomy_cache Application.get_env(:td_dd, :taxonomy_cache)
 
@@ -47,8 +48,8 @@ defmodule TdDdWeb.MetadataController do
 
     start_time = DateTime.utc_now()
 
-    structure_recs = params |> Map.get("data_structures") |> parse_data_structures
     field_recs = params |> Map.get("data_fields") |> parse_data_fields
+    structure_recs = params |> Map.get("data_structures") |> parse_data_structures
 
     relation_recs =
       params |> Map.get("data_structure_relations") |> parse_data_structure_relations
@@ -58,6 +59,8 @@ defmodule TdDdWeb.MetadataController do
     end_time = DateTime.utc_now()
 
     Logger.info("Metadata uploaded. Elapsed seconds: #{DateTime.diff(end_time, start_time)}")
+
+    IndexWorker.reindex()
   end
 
   defp parse_data_structures(%Upload{path: path}) do
@@ -105,7 +108,19 @@ defmodule TdDdWeb.MetadataController do
     record
     |> add_metadata(@data_field_modifiable_fields)
     |> to_map(@data_field_keys)
+    |> blanks_to_nils
   end
+
+  defp blanks_to_nils(map) do
+    map
+    |> Enum.map(&blank_to_nil/1)
+    |> Map.new()
+  end
+
+  defp blank_to_nil({:metadata, v}), do: {:metadata, blanks_to_nils(v)}
+  defp blank_to_nil({k, v}), do: {k, blank_to_nil(v)}
+  defp blank_to_nil(""), do: nil
+  defp blank_to_nil(v), do: v
 
   defp csv_to_relation(record) do
     record