+++ priv/repo/migrations/20181220141635_delete_duplicate_fields.exs
@@ -1,0 +1,13 @@
+defmodule TdDd.Repo.Migrations.DeleteDuplicateFields do
+  use Ecto.Migration
+
+  def up do
+    execute("""
+    delete from data_fields where id not in (select data_field_id from versions_fields)
+    """)
+  end
+
+  def down do
+    
+  end
+end